Added methods for divining type information.
This is building towards being able to verify type correctness in the
AST.  I'm using GLSL as a reference, though I'm not sure how portable
the assumptions are for that.

This provides methods for determining if something is a scalar, a
vector, or a matrix.  These categories are considered to be mutually
exclusive for now (I'm not expecting it to be useful for something to be
both a vector and a matrix right now).

Also, there is a "vector-size" method which will take a value or a list
and determine the total size of the element or elements.  Eg, a scalar
value has a "vector-size" of 1; a vector can be 2, 3, or 4; and a matrix
can be 4; 9; or 16.  (vector-size '(1 2 #f #:vec3 #:mat3)) evaluates to
15.

Path lookups are also valid for all of these, so for the following
struct:

(struct* 'contrived
  '(some-scalar 10)
  '(some-vector #:vec4)
  '(some-matrix #:mat4))

The following evaluate true:

(scalar? '(contrived some-scalar))
(vector? '(contrived some-vector))
(matrix? '(contrived some-matrix))

The "vector-size" method also accepts path lookups.

I expect that this will be useful for determining correctness in GLSL's
casting conventions.  For example, you can construct a matrix with any
combination of scalars and vectors so long as the total size is correct.