Code for describing and introspecting shader structures.
This commit pulls out ```action-struct``` into shader-struct.scm.
Currently this serves to sort the fields defined on the struct into
three categories, so that they can be introspected more easily later on.
The categories are:

 - Buffers, which are either going to represent attribute variables or
   varrying variables, depending on if the variable is being accessed
   from a vertex shader or fragment shader.

 - Controls, which are usually probably always going to be uniform
   variables except that if the struct is being instanced, in which case
   a control might be promoted to a buffer.

 - Functions, which are functions associated to the struct, but also
   aliases and constants.  These can reference buffers and control and
   other functions on the struct like local variables.  A lambda can be
   used to access other namespaces.

This commit also includes ```print-struct```, which is a handy dandy way
of printing out the type data for a given struct.

An area where this can be improved further would be by adding a
validation step of some kind to ```action-struct```, to run after the
fields have been split into categories.  With the excetpion of lambda
structs, there should be enough information here for this to be viable
to implement now.

If a type assertion form were added, then it would be possible to
validate the lambdas, too, though it would make sense to wait until
later before adding type validation.