Implementations for action-shader, action-program, and some validation.
This commit adds implementations for action-shader and action-program,
replacing the pass-through stand-ins I had before.  This uses an
associative list style data structure for each like is used with
action-struct.

I think this could be simplified a bit:

 - Replace define-vertex-shader etc with define-shader.

 - Change define-program syntax to use an a-list for pairing shader
   stages to their implementations.

I don't think this would change much for how action-shader is
implemented, but it would simplify action-program a little.

The usage of (cdr (assoc etc etc)) could probably be cleaned up by
replacing it with a method that does that but also just returns '() when
the record is missing.  There might be a built in method that does this,
though so idk.

As for the validation part of this commit, this basically checks to make
sure the shader inputs don't conflict, and that the 'interpolates' have
a corresponding input of a matching name and type.

Some places to improve going forward:

 - Rename define-program to define-gpu-program?

 - It might be useful to cache the validated input results on the
   program object, as an intermediary to the munging needed to produce
   the final definitions, as well as to make it easier to generate
   frontend boilerplate bindings.

 - From what I have so far, validation should be possible now (though it
   might be best to put it off until I'm more sure about the finality of
   the syntax so far).  At the point the program object is populated,
   all of the information it needs to be validated should be present in
   the environment object.