Revised shader syntax, refactored program action.
Now there is just one special form for defining a shader,
define-shader-stage.  The type of the shader is then determined by the
define-shader-program special form.  Said define-shader-program now
takes an associative list as its input instead of a list of shaders.

The validation rules for when to use things like the interpolate input
are still the same, they are just applied when combining the shader
program now.

The contents of the program object is more robust now.  It includes the
following fields:

 - shaders, which is the a-list supplied by the user
 - types, which will usually just be (#:vertex #:fragment)
 - inputs, which are the combined inputs from the referenced shaders
 - transports, which are the transports from the fragment shader