Simplifying gather-program-vars.
This change removes part of the quasi-environment-model system for
tracking bindings, reducing it to just a single frame of bindings that
is passed around.

Later, when this is able to recurse into field definitions on the models
themselves, this will just generate new bindigs to pass into the
recursive calls.  For this, there is global scope, and are no closures.

This change also removes the extra quote from return values, ensuring
that they look like so: (<model> screen-position).  I've done so by just
removing the quote from the syntax for the record lookups entirely.