gather-program-vars can now traverse into records
The gather-program-vars method can now traverse into records, and read
local references to other fields in that record.

This is almost complete, with these remaining cases:

 1. Passing a binding name into a function call.

 2. Invoking a local reference as a function call.

 3. Aliases, via let or define.

The first two can probably be (badly?) approximated by some variety of
look-ahead mechanism.

With the current shader.scm example, this code as-is only can give us
the following:

(<model> screen-position)
(<model> world-matrix)
(<model> vertex-position)
(<model> texture-map)
(<model> vertex-uv))

What is absent is the <camera> references, which should have resulted
from (<model> screen-position).

But also, this scheme totally does not account for arrays of models, and
I'm not yet sure how to account for that.

It feels like this approach may be a dead-end, and that instead this
should be replaced with some kind of variant on a proper meta-circular
evaluator.