Added inlining function to be used for later type validation.
This commit adds the "inlined-program" method, which reduces a shader
program to a nested list of types and function symbols.

For example, the vertex shader defined in shader.scm will reduce to
this:

  (* (* #:matrix4 #:matrix4) #:matrix4 #:float3)

At this point, it should be fairly easy to start integrating the type
inference system described in eccc11abd2ff4a5be180f3f7a8282e7f1af040ae.

A few notes on going forward:

 - The source mapping stuff that guile provides is probably missing from
   the inlined tree?  I haven't checked, but I would not be surprised.

 - This should probably be pulled into the validator.

 - The validator is probably going to end up being a submodule.

 - The "gather-program-vars" method is likely redundant to this new
   method, and could probably be accomplished by a callback or something
   added to the traversal method in "inlined-program".