Connected type inference demo to inlining output.
I noticed that the example program was totally valid with
"(world-position (* (world-matrix) (vertex-position)))".  I'm not sure
how, but I guess that is fine for now - the type inference system is
pretty basic right now.

Other cleanup tasks aside, one thing that very definitely needs to be
refactored sooner than later is not using a global state for building
the type inference constraint system, so that several shaders can be
validated in isolation from one another.

A happy moment here is that (* #matrix4 #float3) is a type error, so
I've updated the shader accordingly for now.  Implicit upcasting from
float3 to float4 might be a nice to have in some future iteration, but
an explicit casting interface is far more important.

I plan on keeping with the convention of allowing a variable number of
arguments here provided that the vector size adds up correctly.  That
might mean a separate constraint template than the permuter that the
example is currently using?  Or maybe it just generates a permutation
matrix for every possible combination that adds up to the target size?
I guess the later is probably more useful, because if you have a case
like (float4 #:float3 1), then it would be able to know that the last
argument is a float and use the number literal thusly.

There is a lot of space to branch off of here, but I think the most
important things in the near term are:

 - encapsulating mutation

 - remove the "include" for accessing the type inference machinery in
   favor of making it part of the yaga module

 - factoring out unecessary complexity (mostly in the existing stuff)

On that later point, it seems that I can probably unify some of the
program traversal stuff into a single function?  But I don't want to
just end up with something that is just overly cluttered.

Once this is cleaned up, there might be enough here to generate a simple
GLSL program!