Added type definitions for builtin shader functions.
There are some things missing, though.  For one, I don't have a
mechanism for variable numbers of arguments, which affects at least
"atan" and variants of "sample".

Would be especially nice to squash the sample-lod and projective-sample
variants of sample down into one function.  Though, maybe having them as
separate functions is a feature here at least.

For some reason, (permutate '(#:bool #:bool-vector)) returns nothing,
causing register! to crash when called for "not", so that one is
commented out for now.

For the variable args thing, maybe this could partially be resolved by
registering them separately.

Also permutating "#:matrix" is useless because the only function that
seems to use it is component-wise-multiply, which only operates on mat2,
mat3, and mat4.

I suspect this type inference system will also be instrumental in
eventually rewriting to GLSL or HLSL - and some of the builtins for Yaga
will need rewrite rules.  EG, I'm planning on using the same not
operator for bools as for bool vectors, but GLSL splits the two into a
syntax operator and a function.

The "any" function might just be "or", which could be really silly
because I don't see why not just support something like this:

  (or #:bool #:bool3 #:bool #:bool2 #:bool)

Which would probably translate to:

  (bool || any(bool3) || bool || any(bool2) || bool)

Something to think about.

One last thing, is, how to specify when it is syntatically ok to chain
things (like in the above "or" example).  I'm kind of tempted to be able
to do it for any function who's output is the same as their inputs.  Or,
any function for which one of the permutations in the permutation matrix
follows that pattern.  For example, being able to chaine min and max
would be neat.  Eg: (min 1 2 3 4).

Min and Max both have a (#:vector #:float) permutation that I'm not sure
what it does off the top of my head?

I guess it could also just be legal to do something like

  (min #:vector #:vector #:float)

so long as all of the vectors had the same size.